refactor decks naming
