+++ src/pages/decks/api/decks-api.ts
@@ -1,14 +1,7 @@
-import { CreateDecksArgs, Deck, DecksArgs, DecksResponse } from '@/pages/decks/api/decks.types'
-import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
+import { CreateDecksArgs, Deck, DecksArgs, DecksResponse } from '@/pages/decks/api/decks-types'
+import { baseAPI } from '@/services/base-api'
 
-export const decksApi = createApi({
-  baseQuery: fetchBaseQuery({
-    baseUrl: 'https://api.flashcards.andrii.es',
-    credentials: 'include',
-    prepareHeaders: headers => {
-      headers.append('x-auth-skip', 'true')
-    },
-  }),
+export const decksApi = baseAPI.injectEndpoints({
   endpoints: builder => {
     return {
       createDecks: builder.mutation<Deck, CreateDecksArgs>({
@@ -16,21 +9,21 @@ export const decksApi = createApi({
         query: body => ({
           body,
           method: 'POST',
-          url: `v1/decks`,
+          url: `/v1/decks`,
         }),
       }),
-      deleteDecks: builder.mutation<Deck, { id: string }>({
+      deleteDecks: builder.mutation<void, { id: string }>({
         invalidatesTags: ['Decks'],
         query: ({ id }) => ({
           method: 'DELETE',
-          url: `v1/decks/${id}`,
+          url: `/v1/decks/${id}`,
         }),
       }),
       getDeckById: builder.query<DecksArgs, { id: string }>({
         providesTags: ['Decks'],
         query: ({ id }) => ({
           method: 'GET',
-          url: `v1/decks/${id}`,
+          url: `/v1/decks/${id}`,
         }),
       }),
       getDecks: builder.query<DecksResponse, DecksArgs | void>({
@@ -38,7 +31,7 @@ export const decksApi = createApi({
         query: args => ({
           method: 'GET',
           params: args ?? undefined,
-          url: `v2/decks`,
+          url: `/v2/decks`,
         }),
       }),
       updateDecks: builder.mutation<Deck, { data: FormData; id: string }>({
@@ -46,13 +39,11 @@ export const decksApi = createApi({
         query: ({ data, id }) => ({
           body: data,
           method: 'PATCH',
-          url: `v1/decks/${id}`,
+          url: `/v1/decks/${id}`,
         }),
       }),
     }
   },
-  reducerPath: 'baseApi',
-  tagTypes: ['Decks'],
 })
 
 export const {
